Changing 'records' counter from int to long to avoid overflow (we're at 2,052,835,068 on production logstorage server)

+++ src/main/java/com/flaptor/indextank/storage/RawLog.java
@@ -40,7 +40,7 @@ public class RawLog {
     private LogRoot root;
 
     private final int segmentSize;
-    private int records;
+    private long records;
     private int segmentRecords;
 
     private long lastTime;
